Add docs to Platform, fix Dokka not generating documentation
Looks like we have two choices. The first choice is to allow Dokka to see our dependencies, which means that no documentation gets generated. Or, we don't allow Dokka to see our dependencies, which means any external types will be unresolved and show as errors in our docs. At least with the latter, the docs actually generate, so I choose the latter.